+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.0.1 - Requires SelfHelp 6.5.0+ (Unpublished)
 ### New Features
- -  rework calendar. Now it is based  on [FullCalendar](https://fullcalendar.io) library
+ - rework calendar. Now it is based  on [FullCalendar](https://fullcalendar.io) library
+ - add new style `calendars`, where the user can set their calendars, later they can be used for entering events on the `calendar` style
 
 # v1.0.0